cloud_function/requirements.txt
Flask
google-api-python-client==1.7.8
